refactored googlevisionAPI class
